requirements.txt
chromadb
crewai
crewai_tools
duckduckgo-search
fastapi
feedparser
langchain
langchain-community
langchain-core
langchain-openai
langchain-ollama
langchainhub
langchain-chroma
matplotlib
numpy
psycopg2-binary
python-dotenv
PyPDF2
requests
scikit-learn
sqlalchemy
uvicorn
google-cloud-firestore
langchain_google_firestore
gpt4all
jupyterlab
llama-cpp-python
PyMuPDF
langchain-google-genai
langgraph
grandalf
langchain_tavily
langchain-groq
langgraph-prebuilt
langgraph-checkpoint
langgraph-checkpoint-sqlite
pytest
sentence-transformers